Finished Saving Board to External File
Finally finished up saving the board to an external file.  It works
really well!  I'm really happy about it.  The editor needs to be
refreshed for the file to show up, but it works.  It works on top of the
system for reading in files, so everything works together very
harmoniously.